pyproject.toml
[project]
name = "roster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.5.2",
    "crispy-bootstrap4>=2024.10",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.1",
    "django-admin-rangefilter>=0.13.2",
    "django-allauth>=65.8.0",
    "django-cors-headers>=4.7.0",
    "django-crispy-forms>=2.4",
    "django-debug-toolbar>=5.2.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "environs[django]>=14.1.1",
    "gunicorn>=23.0.0",
    "ortools>=9.12.4544",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.0",
    "python-dateutil>=2.9.0.post0",
    "watchdog[watchmedo]>=6.0.0",
]
